closes #17 - GML data displayed properly
There was a problem when displaying GML data,
because it is not uncommon that while formatting
the data inside the xml some tabs and line
breaks are included. This made the data unrea-
dable in NavTable until you edit it.
A trim is now made when rendering the data
strings (the value itself remains the same).
Also fixed some encoding problems which I
committed by mistake.