closes #48 CUD operations don't work without View
If you close the View from where you opened NavTable,
any change made on the layer from Navtable will re-
sult in an Exception because it looks for the View
in order to update it. Now, instead of searching for
the View when saving, NavTableExtension gets the
MapControl from it when you open NavTable, and it's
stored in an attribute. ToggleEditing and NavTable
have been modified for handling this new element.
Also did some minor code refactoring and cleaning.