closes #48 CUD operations don't work without View
If you close the View from where you opened NavTable,
any change made on the layer from Navtable will re-
sult in an Exception because it looks for the View
in order to update it. ToggleEditing and NavTable
have been modified for including the option to save
without having to update a View.