In AlphanumericNavtable the SelectableDataSource instance is got from the IEditableSource passed in the constructor through the overrided method getRecordset.
In AbstractForm, Navtable and AbstractNavtable the SelectableDataSource
instance is got from the FLyVect layer passed in the constructor through the
method getRecordset in AbstractNavtable

There is no field for the SelectableDataSource instance in any class, so there is
no need to pass it to parent classes in Alphanumeric.

This commit deprecates the parent constructors that receive a SelectableDataSource
parameter as there are not needed.