Try to parse the value if the db type is unknown.
When saving the input data, if the DB column type is
OTHER (which usually means unknown), we try to parse
the text as a number. And if it's a number, we try
first as an int, and if we lose information, then re-
turn it as a double.
If any exception is thrown, we just return the original
string as we already did.